Added IRC to index.html
Adds a title and an iframe containing a kiwi IRC client pre-configured to launch a connection to webchat.freenode.net#columbus-elixir.

End users need only enter a nick-name and verify their humanity prior to joining.

Currently, the #columbus-elixir channel is set up with a secure password that I have access to. Likewise for the @columbus-elixir registered account on freenode.